Split food purchases into schema file and per-year files
One use case is that it makes it easier to scroll to the bottom of the
file when accessing it on my phone, which I sometimes need to do when
I want to check on recent purchases while shopping at Trader Joe's.

More broadly, this helps keep the file size bounded over time, rather
than having it grow linearly.